pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yoda"
version = "0.1.0"
description = "Yoda — a personal AI assistant with infinite memory, tool use, and knowledge graph reasoning"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Yoda Team" }]

dependencies = [
    "pydantic>=2.6,<3",
    "pyyaml>=6.0,<7",
    "tiktoken>=0.7,<1",
    "httpx>=0.27,<1",
    "anthropic>=0.40,<1",
    "openai>=1.50,<2",
    "rich>=13.7,<14",
    "click>=8.1,<9",
    "chromadb>=0.5,<1",
    "sentence-transformers>=3.0,<4",
    "numpy>=1.26,<3",
    "networkx>=3.2,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5",
    "mypy>=1.11",
]
local = [
    "ollama>=0.4,<1",
]
faiss = [
    "faiss-cpu>=1.8,<2",
]
computer = [
    "pyautogui>=0.9,<1",
    "Pillow>=10.0,<12",
    "pytesseract>=0.3,<1",
    "pyperclip>=1.8,<2",
]
ocr = [
    "easyocr>=1.7,<2",
]

[project.scripts]
yoda = "yoda.cli:main"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
